# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "使用GO语言gin框架开发的Restful API example",
        "title": "Restful API",
        "contact": {
            "name": "lughong",
            "url": "https://github.com/lughong/gin-api-demo",
            "email": "1586668924@qq.com"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "登录成功返回一个token，后面访问操作都需要带上这个token值作校验",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"msg\":\"OK\",\"data\":{\"token\":\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE1Nzg5OTU0NTMsImlkIjozLCJuYmYiOjE1Nzg5OTU0NTMsInVzZXJuYW1lIjoibGlzaSJ9.agmaafda4LwOqkqDbIkpV9AHkdaoFVHhOMkasu_qCTM\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "成功后，返回新增用户名称",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "新增一个用户",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "age",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"msg\":\"OK\",\"data\":{\"username\":\"zhangsan\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/{username}": {
            "get": {
                "description": "从数据库中获取用户信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"msg\":\"OK\",\"data\":{\"username\":\"zhangsan\",\"age\":18}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}